+++ protoc_plugin/lib/names.dart
@@ -485,28 +485,39 @@ bool _isDartFieldName(String name) => name.startsWith(_dartFieldNameExpr);
 final _dartFieldNameExpr = RegExp(r'^[a-z]\w+$');
 
 /// Names that would collide as top-level identifiers.
-final List<String> forbiddenTopLevelNames = <String>[
+const forbiddenTopLevelNames = <String>[
   'List',
   'Function',
   'Map',
   ..._dartReservedWords,
 ];
 
-final List<String> reservedMemberNames = <String>[
+const reservedMemberNames = <String>[
   ..._dartReservedWords,
   ...GeneratedMessage_reservedNames,
   ..._generatedMessageNames
 ];
 
-final List<String> forbiddenExtensionNames = <String>[
+const forbiddenExtensionNames = <String>[
   ..._dartReservedWords,
   ...GeneratedMessage_reservedNames,
   ..._generatedMessageNames
+];
+
+const serviceReservedMemberNames = <String>[
+  ..._dartReservedWords,
+  ..._serviceNames,
+];
+
+const _serviceNames = <String>[
+  // From GeneratedService
+  r'createRequest',
+  r'handleCall',
 ];
 
 // List of Dart language reserved words in names which cannot be used in a
 // subclass of GeneratedMessage.
-const List<String> _dartReservedWords = [
+const _dartReservedWords = <String>[
   'assert',
   'bool',
   'break',