+++ protoc_plugin/test/generated_message_test.dart
@@ -17,11 +17,40 @@ import '../out/protos/package1.pb.dart' as p1;
 import '../out/protos/package2.pb.dart' as p2;
 import '../out/protos/package3.pb.dart' as p3;
 import '../out/protos/reserved_names.pb.dart';
+import '../out/protos/reserved_names.pbserver.dart';
 import '../out/protos/reserved_names_extension.pb.dart';
 import '../out/protos/reserved_names_message.pb.dart';
 import '../out/protos/toplevel.pb.dart';
 import '../out/protos/toplevel_import.pb.dart' as t;
 import 'test_util.dart';
+
+class ReservedNamesService extends ReservedNamesServiceBase {
+  @override
+  Future<SimpleResponse> abc(ServerContext ctx, SimpleReq request) async =>
+      SimpleResponse(test: 'Abc called');
+
+  @override
+  Future<SimpleResponse> abc_Pre_(ServerContext ctx, SimpleReq request) async =>
+      SimpleResponse(test: 'abc_Pre called');
+
+  @override
+  Future<SimpleResponse> handleCall_(
+          ServerContext ctx, SimpleReq request) async =>
+      SimpleResponse(test: 'handleCall called');
+
+  @override
+  Future<SimpleResponse> if_(ServerContext ctx, SimpleReq request) async =>
+      SimpleResponse(test: 'If called');
+
+  @override
+  Future<SimpleResponse> new_(ServerContext ctx, SimpleReq request) async =>
+      SimpleResponse(test: 'New called');
+
+  @override
+  Future<SimpleResponse> x32SillyName_(
+          ServerContext ctx, SimpleReq request) async =>
+      SimpleResponse(test: '_32SillyName called');
+}
 
 void main() {
   final throwsInvalidProtocolBufferException =
@@ -966,5 +995,22 @@ void main() {
     );
 
     assertAllFieldsSet(value);
+  });
+
+  test('Service name disambiguation', () async {
+    final service = ReservedNamesService();
+    expect(await service.handleCall(ServerContext(), 'New', SimpleReq()),
+        SimpleResponse(test: 'New called'));
+    expect(await service.handleCall(ServerContext(), 'If', SimpleReq()),
+        SimpleResponse(test: 'If called'));
+    expect(await service.handleCall(ServerContext(), 'Abc', SimpleReq()),
+        SimpleResponse(test: 'Abc called'));
+    expect(await service.handleCall(ServerContext(), 'abc_Pre', SimpleReq()),
+        SimpleResponse(test: 'abc_Pre called'));
+    expect(
+        await service.handleCall(ServerContext(), '_32SillyName', SimpleReq()),
+        SimpleResponse(test: '_32SillyName called'));
+    expect(await service.handleCall(ServerContext(), 'handleCall', SimpleReq()),
+        SimpleResponse(test: 'handleCall called'));
   });
 }